+++ datatig-cli.py
@@ -1,4 +1,0 @@
-from datatig.cli import main
-
-if __name__ == "__main__":
-    main()