+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+* Removed `datatig-cli.py` - call via `python -m datatig.cli` instead 
+
 ## [0.1.1] - 2021-04-15
 
 ### Fixed