+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-07-22
+
 ### Changed
 
 * Removed `datatig-cli.py` - call via `python -m datatig.cli` instead